+++ node.js
@@ -1,6 +1,5 @@
 import WseClient  from './src/client.js'
 import WseServer  from './src/server.js'
-import WseMServer from './src/server-m.js'
 import WSE_REASON from './src/reason.js'
 
-export { WseClient, WseServer, WseMServer, WSE_REASON }
+export { WseClient, WseServer, WSE_REASON }
